adding security

+++ server/routes/convo.router.js
@@ -54,7 +54,7 @@ router.get('/:id', (req, res) => {
 router.delete('/:id', (req, res) => {
   if (req.isAuthenticated()) {
     const deleteInfo = req.params.id;
-    pool.query('DELETE FROM "comments" WHERE "id"=$1;', [deleteInfo])
+    pool.query('DELETE FROM "comments" WHERE "id"=$1 AND "user_id" = $2;', [deleteInfo, req.user.id])
       .then((result) => {
         res.sendStatus(200);
       }).catch((error) => {
@@ -73,10 +73,12 @@ router.put('/:id', (req, res) => {
     const newComment = req.body; 
     const queryText = `UPDATE comments
                       SET "comment" = $1 
-                      WHERE id=$2;`;
+                      WHERE id=$2
+                      AND "user_id" = $3;`;
     const queryValues = [
       newComment.comment,
       req.params.id,
+      req.user.id,
     ];
     pool.query(queryText, queryValues)
       .then(() => { res.sendStatus(200); })
